Anexos

+++ src/main/java/mz/fipag/grm/dao/DocRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import mz.fipag.grm.domain.Doc;
 
-public interface DocRepository extends JpaRepository<Doc, Integer>{
+public interface DocRepository extends JpaRepository<Doc, Long>{
 
 }
